refactor(host/parents-gateway): rename remote view files for consistency
Match the existing host naming convention (HomeView, StudentsView) by
naming container views after the feature, not the underlying remote's
product name: GroupsGatewayView -> GroupsView. Also rename the shared
remote-loading helper file to match its exported function name,
consistent with how hooks are named elsewhere in this codebase:
parents-gateway-remote.tsx -> createParentsGatewayRemoteView.tsx.
