+++ apps/host/src/containers/GroupsView.tsx
@@ -1,0 +1,3 @@
+import { createParentsGatewayRemoteView } from '~/containers/createParentsGatewayRemoteView';
+
+export const GroupsView = createParentsGatewayRemoteView('pg/Groups', 'Groups');
